\name{interiorPoint}
\alias{interiorPoint}
\title{
Find an interior point of a polyhedron.
}
\description{
Finds an interior point in the augmented question matrix X and answer vector a.
}
\usage{
interiorPoint(X, a)
}
\arguments{
  \item{X}{
Augmented matrix describing the polyhedron/conjoint cards seend
}
  \item{a}{
User responses to each of the conjoint cards.
}
}
\value{
Returns a vector describing an interior point inside the system of equations described by X and a.
}
\examples{
}
