\name{rescale}
\alias{rescale}
\title{
Rescales a vector
}
\description{
Rescales an approximate solution to the center of a polyhedron
}
\usage{
rescale(vec)
}
\arguments{
  \item{vec}{
Vector describing the approximate center of the polyehdron
}
}
\value{
Returns an approximation of the center that has been rescaled
}
\examples{
}
