\name{setLevels}
\alias{setLevels}
\title{
Reshapes a vector.
}
\description{
Reshapes the a vector to match attribute level preferences
}
\usage{
setLevels(originalVector, pref)
}
\arguments{
  \item{originalVector}{ Vector to be reshaped}
  \item{pref}{Attribute level preferences.  }
}

\value{
Returns originalVector reshaped to match non-zero levels of pref.
}

